code/contract-agent-eval/pyproject.toml
[project]
name = "contract-agent-eval"
version = "0.1.0"
description = "Contract Agent Evaluation with Opik tracing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "neo4j-rust-ext>=6.0.2.0",
    "opik>=1.8.96",
    "python-dotenv>=1.0.0",
]

